JC-907 Fixed problem
	modified:   jcommune-model/src/main/java/org/jtalks/jcommune/model/dao/SectionDao.java
	modified:   jcommune-model/src/main/java/org/jtalks/jcommune/model/dao/hibernate/SectionHibernateDao.java
	modified:   jcommune-model/src/main/resources/org/jtalks/jcommune/model/entity/Section.hbm.xml
	modified:   jcommune-model/src/test/java/org/jtalks/jcommune/model/ObjectsFactory.java
	modified:   jcommune-model/src/test/java/org/jtalks/jcommune/model/dao/hibernate/SectionHibernateDaoTest.java
	modified:   jcommune-service/src/test/java/org/jtalks/jcommune/service/transactional/TransactionalSectionServiceTest.java
	modified:   jcommune-view/jcommune-web-controller/src/main/java/org/jtalks/jcommune/web/controller/SectionController.java
	modified:   jcommune-view/jcommune-web-controller/src/test/java/org/jtalks/jcommune/web/controller/SectionControllerTest.java
	modified:   jcommune-view/jcommune-web-view/src/main/resources/org/jtalks/jcommune/model/datasource.properties

+++ jcommune-model/src/main/java/org/jtalks/jcommune/model/dao/SectionDao.java
@@ -42,5 +42,5 @@ public interface SectionDao extends ParentRepository<Section> {
      * @param branches branches of current section
      * @return count branches
      */
-    int getCountAvailableBranches(JCUser user, List<Branch> branches);
+    Long getCountAvailableBranches(JCUser user, List<Branch> branches);
 }
